refactor: rename AppRealization to AppRouting
