+++ src/components/AppRouting/index.js
@@ -1,0 +1,4 @@
+import AppRoutingCC from '@components/AppRouting/AppRoutingCC';
+import AppRoutingFC from '@components/AppRouting/AppRoutingFC';
+
+export { AppRoutingCC, AppRoutingFC };
